escet_backend/requirements.txt
# Django Core
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==24.1

# Database
psycopg2-binary==2.9.9  # PostgreSQL adapter (optional)

# Authentication
djangorestframework-simplejwt==5.3.1

# Image Processing
Pillow==10.2.0

# Environment Variables
python-decouple==3.8

# Utilities
python-dateutil==2.8.2

# Development
django-debug-toolbar==4.2.0
